feat: implement networks management UI with create, edit, and delete functionalities
