+++ src/app/admin/networks/_components/DeleteNetworkDialog.tsx
@@ -1,0 +1,56 @@
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+import type { Network } from "@/server/networks/types";
+
+interface DeleteNetworkDialogProps {
+  open: boolean;
+  target: Network | null;
+  isDeleting: boolean;
+  onOpenChange: (open: boolean) => void;
+  onConfirmDelete: () => void;
+}
+
+export function DeleteNetworkDialog({
+  open,
+  target,
+  isDeleting,
+  onOpenChange,
+  onConfirmDelete,
+}: DeleteNetworkDialogProps) {
+  return (
+    <AlertDialog onOpenChange={onOpenChange} open={open}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Delete network?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This will permanently delete{" "}
+            <span className="font-medium text-foreground">
+              {target?.organization}
+            </span>
+            .
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={isDeleting}>Cancel</AlertDialogCancel>
+          <AlertDialogAction
+            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+            disabled={isDeleting}
+            onClick={() => {
+              void onConfirmDelete();
+            }}
+          >
+            {isDeleting ? "Deleting..." : "Delete"}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
